swarm: remove int64 cast from payment/disconnect thresholds)

+++ swarm.go
@@ -123,8 +123,8 @@ func NewSwarm(config *api.Config, mockStore *mock.NodeStore) (self *Swarm, err e
 		}
 		swapParams := &swap.Params{
 			LogPath:             self.config.SwapLogPath,
-			DisconnectThreshold: int64(self.config.SwapDisconnectThreshold),
-			PaymentThreshold:    int64(self.config.SwapPaymentThreshold),
+			DisconnectThreshold: self.config.SwapDisconnectThreshold,
+			PaymentThreshold:    self.config.SwapPaymentThreshold,
 		}
 
 		// create the accounting objects
